Rid of TypeResolver.ResolveType() call from Process_Types_Ps1Xml()
