+++ src/System.Management.Automation/engine/InitialSessionState.cs
@@ -3929,13 +3929,6 @@ namespace System.Management.Automation.Runspaces
 
             this.Assemblies.Add(assemblyEntry);
 
-            // entry from types.ps1xml references a type (Microsoft.PowerShell.Commands.SecurityDescriptorCommandsBase) in this assembly
-            if (psSnapInInfo.Name.Equals(CoreSnapin, StringComparison.OrdinalIgnoreCase))
-            {
-                assemblyEntry = new SessionStateAssemblyEntry("Microsoft.PowerShell.Security", null);
-                this.Assemblies.Add(assemblyEntry);
-            }
-
             if (cmdlets != null)
             {
                 foreach (SessionStateCmdletEntry cmdlet in cmdlets.Values)
